integrated dgi algorithm
